requirements.txt
aiomysql==0.2.0
alembic==1.12.0
anyio==3.7.1
astroid==3.0.0
async-timeout==4.0.3
Cerberus==1.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
coverage==7.3.2
crontab==1.0.1
cryptography==41.0.4
dill==0.3.7
docker==7.0.0
exceptiongroup==1.1.3
fakeredis==2.20.0
fastapi==0.104.0
freezegun==1.2.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
hvac==1.2.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mock==5.1.0
packaging==23.2
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.19.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.9.2
pyhcl==0.4.5
PyJWT==2.8.0
pylint==3.0.0
PyMySQL==1.1.0
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-env==1.1.3
python-dateutil==2.8.2
python-ldap==3.4.4
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rq==1.15.1
rq-dashboard==0.6.7.2
rq-exporter==2.1.2
rq-scheduler==0.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
